+++ adminer/include/bootstrap.inc.php
@@ -44,7 +44,11 @@ define('Adminer\HTTPS', ($_SERVER["HTTPS"] && strcasecmp($_SERVER["HTTPS"], "off
 ini_set("session.use_trans_sid", '0'); // protect links in export
 ini_set("arg_separator.output", "&"); // some hosts set it to "&amp;" which would break http_build_query()
 // arg_separator.input is not checked - it is PHP_INI_PERDIR so we couldn't fix it and a value without & would break almost every PHP application
-if (!defined("SID")) {
+define('Adminer\SESSION_NAME', session_name()); // PHP stores the upload progress in a session named by php.ini, session_name() below changes this value
+if (isset($_GET["upload"])) {
+	include DIR . "upload.inc.php";
+}
+if (function_exists('session_status') ? session_status() == PHP_SESSION_NONE : !defined("SID")) { // session_status() available since PHP 5.4
 	session_cache_limiter(""); // to allow restarting session
 	session_name("adminer_sid"); // use specific session name to get own namespace
 	// ini_set() may be disabled